feat: :sparkles: move posts index from index page
move to posts add redirect for garden > posts
