+++ src/routes/[yyyy]/[mm]/[dd]/[slug]/index.svelte
@@ -1,3 +1,8 @@
+<!-- 
+  Redirects from my old blog filing structure yyyy/mm/dd/post-title
+  to posts/post-title thanks to rodneylab for the example 👇
+  https://github.com/rodneylab/sveltekit-blog-mdx/blob/dev__redirect/src/routes/%5Byear%5D/%5Bmonth%5D/%5Bday%5D/%5Bslug%5D/index.svelte
+ -->
 <script context="module">
   export async function load({ page }) {
     const { slug } = page.params
