+++ src/routes/garden/index.svelte
@@ -1,0 +1,8 @@
+<script context="module">
+  export async function load() {
+    return {
+      status: 301,
+      redirect: `/posts/`,
+    }
+  }
+</script>
